requirements.txt
Django>=4.2
djangorestframework>=3.14
django-ckeditor>=6.3.0
Pillow>=10.0.0
